refactor(tailwind): move InstallPrompt to tailwind
