--- src/components/InstallPrompt/style.module.css
@@ -1,150 +1,0 @@
-/*
-License:
-
-This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
-the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
-*/
-
-.container {
-  position: fixed;
-  position: sticky;
-  bottom: 0px;
-  left: 0px;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  color: #111827;
-  color: var(--soft-text-color);
-  background-color: white;
-  background-color: var(--main-bg-color);
-  border-style: solid;
-  border-color: #e5e8f4;
-  border-color: var(--dialog-border-color);
-  border-width: 0px;
-  border-top-width: 1px;
-  box-shadow: 0 0 2rem rgba(0, 0, 0, 0.25);
-  box-shadow: 0 0 2rem var(--dialog-shadow-color);
-}
-
-.container_initial {
-  composes: container;
-  scale: 0.99;
-  translate: 0 1rem;
-  opacity: 0;
-}
-
-@media (min-width: 640px) {
-  .container {
-    bottom: 0.25rem;
-    left: 0.25rem;
-    border-width: 1px;
-    border-radius: 0.25rem;
-    margin-bottom: env(safe-area-inset-bottom);
-    margin-left: env(safe-area-inset-left);
-    max-width: 24rem;
-  }
-}
-
-.content {
-  padding: 1.25rem;
-}
-
-.buttons_row {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-end;
-  margin: 0.75rem;
-  margin-right: 1.5rem;
-}
-.buttons_row > button + button {
-  margin-left: 1.5rem;
-}
-
-.ios_install {
-  position: fixed;
-  position: sticky;
-  bottom: 0px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100vw;
-  background-color: white;
-  background-color: var(--main-bg-color);
-  border-style: solid;
-  border-color: #e5e8f4;
-  border-color: var(--dialog-border-color);
-  border-width: 0px;
-  border-top-width: 1px;
-  box-shadow: 0 0 2rem rgba(0, 0, 0, 0.25);
-  box-shadow: 0 0 2rem var(--dialog-shadow-color);
-  padding-left: 0.5rem;
-  padding-left: env(safe-area-inset-left, 0.5rem);
-  padding-right: 0.5rem;
-  padding-right: env(safe-area-inset-right, 0.5rem);
-  padding-bottom: env(safe-area-inset-bottom);
-  z-index: 10;
-}
-
-.ios_text {
-  margin: 0.25rem 3rem;
-  margin-left: calc(4rem + env(safe-area-inset-right, 0px));
-  margin-right: calc(4rem + env(safe-area-inset-right, 0px));
-  font-size: 0.875rem;
-  text-align: center;
-  color: #111827;
-  color: var(--soft-text-color);
-}
-
-.ios_icons_row {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 0.25rem;
-  padding-inline: 3px;
-}
-
-.ios_icon {
-  height: 4.9rem;
-  border-radius: 0.25rem;
-  object-fit: contain;
-  object-position: center;
-  padding: 0.25rem;
-  border-radius: 0.25rem;
-  box-shadow: 0 0 2rem var(--dialog-shadow-color);
-}
-
-.ios_icon_space {
-  display: block;
-  margin: auto 0.75rem;
-}
-
-.close_icon {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  right: env(safe-area-inset-right);
-  margin-left: 0.25rem;
-  background-color: transparent;
-  border: 0px;
-}
-.close_icon_svg {
-  width: 3rem;
-  height: 3rem;
-}
-
-@media (min-width: 640px) {
-  .ios_install {
-    bottom: 0.25rem;
-    left: 0.25rem;
-    border-width: 1px;
-    border-radius: 0.25rem;
-    margin-bottom: env(safe-area-inset-bottom);
-    margin-left: env(safe-area-inset-left);
-    max-width: 28rem;
-  }
-
-  .close_icon {
-    right: -5px;
-  }
-}
